+++ symphony/bdk/core/config/model/bdk_datafeed_config.py
@@ -23,8 +23,7 @@ class BdkDatafeedConfig:
         if config is not None:
             self.id_file_path = Path(config.get(DF_ID_FILE_PATH)) if DF_ID_FILE_PATH in config else ""
             self.version = config.get(VERSION)
-            if "retry" in config:
-                self.retry = BdkRetryConfig(config.get("retry"))
+            self.retry = BdkRetryConfig(config.get("retry")) if "retry" in config else None
 
     def get_id_file_path(self) -> Path:
         """
