+++ symphony/bdk/core/service/message/message_service.py
@@ -1,6 +1,7 @@
 from symphony.bdk.core.auth.auth_session import AuthSession
 from symphony.bdk.core.config.model.bdk_retry_config import BdkRetryConfig
 from symphony.bdk.core.service.message.multi_attachments_messages_api import MultiAttachmentsMessagesApi
+from symphony.bdk.gen import ApiException
 from symphony.bdk.gen.agent_api.attachments_api import AttachmentsApi
 from symphony.bdk.gen.agent_model.v4_import_response import V4ImportResponse
 from symphony.bdk.gen.agent_model.v4_imported_message import V4ImportedMessage
@@ -20,7 +21,7 @@ from symphony.bdk.gen.pod_model.message_suppression_response import MessageSuppr
 from symphony.bdk.gen.pod_model.stream_attachment_item import StreamAttachmentItem
 
 from symphony.bdk.core.retry import retry
-
+from symphony.bdk.core.retry.startegy import refresh_session_if_unauthorized
 
 class OboMessageService:
     """Class exposing OBO enabled endpoints for message management, e.g. send a message."""
@@ -33,7 +34,7 @@ class OboMessageService:
         self._auth_session = auth_session
         self._retry_config = retry_config
 
-    @retry
+    @retry(retry=refresh_session_if_unauthorized)
     async def send_message(
             self,
             stream_id: str,
@@ -73,7 +74,19 @@ class OboMessageService:
 
         return await self._messages_api.v4_stream_sid_multi_attachment_message_create_post(**params)
 
-    @retry
+    @retry(retry=refresh_session_if_unauthorized)
+    async def execute_and_retry(self, method, *args):
+        try:
+            return await self.method(*args)
+        except Exception as exc:
+            # handle recovery
+            raise exc
+
+
+
+
+
+    @retry(retry=refresh_session_if_unauthorized)
     async def suppress_message(
             self,
             message_id: str
@@ -114,7 +127,7 @@ class MessageService(OboMessageService):
         self._attachment_api = attachment_api
         self._default_api = default_api
 
-    @retry
+    @retry(retry=refresh_session_if_unauthorized)
     async def list_messages(
             self,
             stream_id: str,
@@ -144,7 +157,7 @@ class MessageService(OboMessageService):
         message_list = await self._messages_api.v4_stream_sid_message_get(**params)
         return message_list.value
 
-    @retry
+    @retry(retry=refresh_session_if_unauthorized)
     async def blast_message(
             self,
             stream_ids: [str],
@@ -184,7 +197,7 @@ class MessageService(OboMessageService):
 
         return await self._messages_api.v4_multi_attachment_message_blast_post(**params)
 
-    @retry
+    @retry(retry=refresh_session_if_unauthorized)
     async def import_messages(
             self,
             messages: [V4ImportedMessage]
@@ -249,7 +262,7 @@ class MessageService(OboMessageService):
         }
         return await self._message_api.v1_message_mid_status_get(**params)
 
-    @retry
+    @retry(retry=refresh_session_if_unauthorized)
     async def get_attachment_types(self) -> [str]:
         """Retrieves a list of supported file extensions for attachments.
         See: `Attachment Types <https://developers.symphony.com/restapi/reference#attachment-types>`_
@@ -263,7 +276,7 @@ class MessageService(OboMessageService):
         type_list = await self._pod_api.v1_files_allowed_types_get(**params)
         return type_list.value
 
-    @retry
+    @retry(retry=refresh_session_if_unauthorized)
     async def get_message(
             self,
             message_id: str
@@ -283,7 +296,7 @@ class MessageService(OboMessageService):
         }
         return await self._messages_api.v1_message_id_get(**params)
 
-    @retry
+    @retry(retry=refresh_session_if_unauthorized)
     async def list_attachments(
             self,
             stream_id: str,
@@ -318,7 +331,7 @@ class MessageService(OboMessageService):
         attachment_list = await self._streams_api.v1_streams_sid_attachments_get(**params)
         return attachment_list.value
 
-    @retry
+    @retry(retry=refresh_session_if_unauthorized)
     async def list_message_receipts(
             self,
             message_id: str
@@ -337,7 +350,7 @@ class MessageService(OboMessageService):
         }
         return await self._default_api.v1_admin_messages_message_id_receipts_get(**params)
 
-    @retry
+    @retry(retry=refresh_session_if_unauthorized)
     async def get_message_relationships(
             self,
             message_id: str
