+++ tests/core/auth/__init__.py
@@ -1,0 +1,7 @@
+from symphony.bdk.core.auth.bot_authenticator import BotAuthenticator
+
+
+class TestBotAuthenticator(BotAuthenticator):
+
+    async def _authenticate_and_get_token(self, api_client):
+        pass
