+++ tests/core/service/connection/connection_service_test.py
@@ -9,7 +9,7 @@ from symphony.bdk.gen.pod_api.connection_api import ConnectionApi
 from symphony.bdk.gen.pod_model.user_connection import UserConnection
 from symphony.bdk.gen.pod_model.user_connection_list import UserConnectionList
 from symphony.bdk.gen.pod_model.user_connection_request import UserConnectionRequest
-from tests.core.retry import minimal_retry_config
+from tests.core.config import minimal_retry_config
 from tests.utils.resource_utils import deserialize_object
