+++ tests/core/service/presence/presence_service_test.py
@@ -12,7 +12,7 @@ from symphony.bdk.gen.pod_model.v2_presence_status import V2PresenceStatus
 from symphony.bdk.gen.pod_model.v2_user_presence import V2UserPresence
 
 from tests.utils.resource_utils import deserialize_object
-from tests.core.retry import minimal_retry_config
+from tests.core.config import minimal_retry_config
 
 
 @pytest.fixture(name="auth_session")
